feat(pager): extract expert-pager-policy leaf crate + moe-pager-driver — RUN-2 unblocked on windows-msvc (#276)
BigMama's blocker: continuum-core cannot compile on x86_64-pc-windows-msvc
(61 unix-socket refs), so the RUN-2 driver could never cargo-run beside
her CUDA serve. Fix per the compression principle: the pure policy moves
to a LEAF crate (std + serde only, windows-clean by constraint documented
in its Cargo.toml) and continuum-core re-exports it — one source of
truth, two consumers, every existing capacity::* path unchanged via
shims (ExpertId re-exported through expert_residency).

New in the leaf beyond the moved modules:
- segment.rs: trace-record parsing (12B LE tkey+e, partial-tail guard
  for the stdio-buffered writer) + streaming TokenSegmenter (her exact
  reappearance boundary rule, (layer,expert) matrix dedup) + TkeyTable
  loader for her tkey-to-layer-matrix.json. 5 tests incl. the torn-tail
  guard and her real JSON shape with the FNV self-check key.
- bin/moe-pager-driver: the RUN-2 loop — tail GGML_MOE_TRACE_FILE from
  offset, segment tokens, feed BanditPlanController, rewrite the plan
  every N tokens with a SMALL ROLLING top-256 pin set (the RUN-1
  lesson), truncation-reset on new serves, per-token glass-box stdout
  (serving hit / chosen decay / per-arm rewards / unknown-tkey count).

Validated against her live fixture end-to-end: 83,968 records segment
to the 1472-expert working set, the bandit visibly adapts (decay 0.99
-> 0.0 -> 0.30 as arms separate), serving hit climbs 0 -> 0.69 warm,
plans rewrite on cadence with 256 pins, zero unknown tkeys.

12 leaf tests + 103 continuum-core capacity tests green; both clippy
gates clean.
