+++ core/continuum-core/Cargo.toml
@@ -44,6 +44,7 @@ path = "src/bin/forge_custodian.rs"
 
 [dependencies]
 tokio.workspace = true
+expert-pager-policy = { path = "../expert-pager-policy" }
 serde.workspace = true
 serde_json.workspace = true
 ron = "0.12"     # RON scene descriptions (Bevy-idiomatic on-disk format)
